+++ src/test/java/com/likelion/mutsasns/controller/api/PostApiControllerTest.java
@@ -1,4 +1,4 @@
-package com.likelion.mutsasns.controller;
+package com.likelion.mutsasns.controller.api;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.likelion.mutsasns.controller.api.PostApiController;
